requirements.txt
Flask==2.2.5
gunicorn==23.0.0
psycopg2-binary==2.9.9
